+++ spec/models/draw_suite_spec.rb
@@ -9,4 +9,19 @@ RSpec.describe DrawSuite do
     it { is_expected.to belong_to(:draw) }
     it { is_expected.to belong_to(:suite) }
   end
+
+  describe 'validations' do
+    it { is_expected.to validate_presence_of(:draw) }
+    it { is_expected.to validate_presence_of(:suite) }
+  end
+
+  describe 'uniqueness constraints' do
+    it 'prevent duplicate draw_suites' do
+      draw = create(:draw)
+      suite = create(:suite)
+      draw.suites << suite
+      expect { draw.suites << suite }.to \
+        raise_error(ActiveRecord::RecordInvalid)
+    end
+  end
 end