+++ spec/models/draw_suites_update_spec.rb
@@ -128,7 +128,7 @@ RSpec.describe DrawSuitesUpdate do
         expect(result[:msg]).to have_key(:error)
       end
 
-      def broken_suites
+      def broken_suites # rubocop:disable Metrics/AbcSize
         klass = 'Suite::ActiveRecord_Associations_CollectionProxy'
         instance_spy(klass).tap do |s|
           allow(s).to receive(:destroy)
@@ -136,6 +136,7 @@ RSpec.describe DrawSuitesUpdate do
           # this is necessary to get through the #find_suites_to_remove private
           # method, which will ideally be refactored eventually
           allow(s).to receive(:available).and_return([Suite.last])
+          allow(s).to receive(:to_ary).and_return([])
         end
       end
     end