+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 * Ensure that suites are properly treated as available after archiving old groups ([#895](https://gitlab.com/yale-sdmp/vesta/issues/895)).
 * Fix the results by student view to only show active draw information ([#897](https://gitlab.com/yale-sdmp/vesta/issues/897)).
 * Ensure that students not already in draws can be individually added to other draws ([#898](https://gitlab.com/yale-sdmp/vesta/issues/898)).
+* Add a uniqueness constraint to draw_suites that prevents duplicate entries ([#899](https://gitlab.com/yale-sdmp/vesta/issues/899)).
 
 ## v2.1.0 - 2019-02-04
 ### Notes for upgrade: